requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.2
django-filter==2.4.0
django-allauth==0.54.0
django-cors-headers==3.14.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
